# docs/openapi.yaml
openapi: 3.1.0
info:
  title: RaghuMayaShop API
  version: 1.0.0
  description: >-
    Multi-tenant SaaS platform API for shop management: auth & 2FA, multi-shop
    memberships, inventory, ledger-first stock, GST billing, customers, finance,
    analytics, subscriptions, referrals, notifications, audit, and platform admin.
    Money is serialized as string; dates are ISO-8601; IDs are UUID v4.
    List endpoints return { data, meta: { page, limit, total } }.
    Errors return { error: { code, message } }.
servers:
  - url: http://localhost:4000/api/v1
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /stock/warehouses:
    post:
      tags: [Stock]
      operationId: createWarehouse
      summary: Create a warehouse (STOCK_ADJUST)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, example: 'Main Godown' }
                code: { type: string, example: GD-01 }
                address: { type: string }
                isDefault: { type: boolean, default: false }
      responses:
        '201':
          description: Warehouse created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Warehouse'
    get:
      tags: [Stock]
      operationId: listWarehouses
      summary: List warehouses (STOCK_VIEW)
      responses:
        '200':
          description: Warehouse list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Warehouse'
  /stock/warehouses/{id}:
    patch:
      tags: [Stock]
      operationId: updateWarehouse
      summary: Update a warehouse (STOCK_ADJUST)
      parameters:
        - $ref: '#/components/parameters/PathId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                code: { type: string }
                address: { type: string }
                isDefault: { type: boolean }
      responses:
        '200':
          description: Warehouse updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Warehouse'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Stock]
      operationId: deleteWarehouse
      summary: Soft delete a warehouse (STOCK_ADJUST)
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: Warehouse deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      deleted: { type: boolean, example: true }
        '404':
          $ref: '#/components/responses/NotFound'
  /stock/in:
    post:
      tags: [Stock]
      operationId: stockIn
      summary: Stock in (STOCK_ADJUST); creates or reuses batch by batchNumber
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StockInRequest'
      responses:
        '201':
          description: Stock added
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      movements:
                        type: array
                        items:
                          $ref: '#/components/schemas/StockMovement'
        '400':
          $ref: '#/components/responses/BadRequest'
  /stock/out:
    post:
      tags: [Stock]
      operationId: stockOut
      summary: Stock out (STOCK_ADJUST); fails on insufficient stock
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StockOutRequest'
      responses:
        '201':
          description: Stock removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      movements:
                        type: array
                        items:
                          $ref: '#/components/schemas/StockMovement'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /stock/alerts/expiry:
    get:
      tags: [Stock]
      operationId: expiryAlerts
      summary: Batches expiring within N days (STOCK_VIEW)
      parameters:
        - name: days
          in: query
          schema: { type: integer, default: 30, example: 30 }
      responses:
        '200':
          description: Expiry alerts
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ExpiryAlert'
...
  /purchases:
    post:
      tags: [Purchases]
      operationId: createPurchase
      summary: Create a purchase (auto stock-in on receiving; idempotent)
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePurchaseRequest'
      responses:
        '201':
          description: Purchase created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Purchase'
    get:
      tags: [Purchases]
      operationId: listPurchases
      summary: List purchases
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: supplierId
          in: query
          schema: { type: string, format: uuid }
        - name: status
          in: query
          schema: { type: string }
        - $ref: '#/components/parameters/FromDate'
        - $ref: '#/components/parameters/ToDate'
      responses:
        '200':
          description: Purchase list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Purchase'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
  /purchases/{id}:
    get:
      tags: [Purchases]
      operationId: getPurchase
      summary: Get purchase details with items
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: Purchase
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Purchase'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Purchases]
      operationId: updatePurchase
      summary: Update purchase header/items before receiving
      parameters:
        - $ref: '#/components/parameters/PathId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePurchaseRequest'
      responses:
        '200':
          description: Purchase updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Purchase'
        '404':
          $ref: '#/components/responses/NotFound'

  # ---------------------------------------------------------------- Billing
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived JWT access token (15 min). Obtain via POST /auth/login or /auth/refresh.
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    Page:
      name: page
      in: query
      schema: { type: integer, default: 1, minimum: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
    FromDate:
      name: fromDate
      in: query
      schema: { type: string, format: date, example: '2026-09-01' }
    ToDate:
      name: toDate
      in: query
      schema: { type: string, format: date, example: '2026-09-30' }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string, format: uuid }
      description: Honored on POST /billing/invoices, /billing/payments, /purchases, /subscriptions/change. Retries with the same key return the original response.
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Conflict (duplicate unique key, insufficient stock, downgrade blocked)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: VALIDATION_ERROR }
            message: { type: string, example: 'quantity must be a positive number' }
    PaginationMeta:
      type: object
      properties:
        page: { type: integer, example: 1 }
        limit: { type: integer, example: 20 }
        total: { type: integer, example: 128 }
...
    StockMovementType:
      type: string
      enum: [IN, OUT, TRANSFER_IN, TRANSFER_OUT, ADJUSTMENT, PURCHASE, SALE]
...
    Warehouse:
      type: object
      properties:
        id: { type: string, format: uuid }
        shopId: { type: string, format: uuid }
        name: { type: string, example: 'Main Godown' }
        code: { type: string, example: GD-01 }
        address: { type: string, nullable: true }
        isDefault: { type: boolean, example: true }
...
    StockMovement:
      type: object
      properties:
        id: { type: string, format: uuid }
        shopId: { type: string, format: uuid }
        warehouseId: { type: string, format: uuid, nullable: true }
        productId: { type: string, format: uuid }
        variantId: { type: string, format: uuid, nullable: true }
        batchId: { type: string, format: uuid, nullable: true }
        movementType: { $ref: '#/components/schemas/StockMovementType' }
        quantity: { type: string, example: '25' }
        unitCost: { type: string, nullable: true, example: '100.00' }
        referenceType: { type: string, nullable: true, example: MANUAL }
        referenceId: { type: string, format: uuid, nullable: true }
        note: { type: string, nullable: true }
        createdById: { type: string, format: uuid, nullable: true }
        createdAt: { type: string, format: date-time }
...
    ExpiryAlert:
      type: object
      properties:
        batchId: { type: string, format: uuid }
        productId: { type: string, format: uuid }
        productName: { type: string, example: 'Premium Rice 25kg' }
        batchNumber: { type: string, example: BATCH-JUN-2026 }
        expiryDate: { type: string, format: date, example: '2026-12-31' }
        daysToExpiry: { type: integer, example: 12 }
        quantity: { type: string, example: '40' }
    StockInRequest:
      type: object
      required: [warehouseId, items]
      properties:
        warehouseId: { type: string, format: uuid }
        referenceType: { type: string, default: MANUAL, example: MANUAL }
        referenceId: { type: string, format: uuid }
        note: { type: string, example: 'Monthly restock' }
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [productId, quantity]
            properties:
              productId: { type: string, format: uuid }
              variantId: { type: string, format: uuid }
              batchNumber: { type: string, example: BATCH-JUN-2026 }
              expiryDate: { type: string, format: date }
              manufacturedDate: { type: string, format: date }
              quantity: { type: string, example: '25' }
              unitCost: { type: string, example: '100.00' }
    StockOutRequest:
      type: object
      required: [warehouseId, items]
      properties:
        warehouseId: { type: string, format: uuid }
        referenceType: { type: string, default: MANUAL }
        referenceId: { type: string, format: uuid }
        note: { type: string }
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [productId, quantity]
            properties:
              productId: { type: string, format: uuid }
              variantId: { type: string, format: uuid }
              batchId: { type: string, format: uuid }
              quantity: { type: string, example: '5' }
...
    PurchaseItem:
      type: object
      properties:
        id: { type: string, format: uuid }
        productId: { type: string, format: uuid }
        productName: { type: string, example: 'Premium Rice 25kg' }
        variantId: { type: string, format: uuid, nullable: true }
        batchNumber: { type: string, nullable: true }
        expiryDate: { type: string, format: date, nullable: true }
        quantity: { type: string, example: '100' }
        unitCost: { type: string, example: '1050.00' }
        lineTotal: { type: string, example: '105000.00' }
    Purchase:
      type: object
      properties:
        id: { type: string, format: uuid }
        shopId: { type: string, format: uuid }
        supplierId: { type: string, format: uuid, nullable: true }
        supplierName: { type: string, nullable: true }
        purchaseNumber: { type: string, example: PO-2026-0042 }
        purchaseDate: { type: string, format: date }
        totalAmount: { type: string, example: '105000.00' }
        status: { type: string, example: RECEIVED }
        notes: { type: string, nullable: true }
        items: { type: array, items: { $ref: '#/components/schemas/PurchaseItem' } }
        createdAt: { type: string, format: date-time }
    CreatePurchaseRequest:
      type: object
      required: [items]
      properties:
        supplierId: { type: string, format: uuid }
        purchaseNumber: { type: string, example: PO-2026-0042 }
        purchaseDate: { type: string, format: date, example: '2026-09-20' }
        warehouseId: { type: string, format: uuid }
        notes: { type: string, example: 'Festival season stock' }
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [productId, quantity, unitCost]
            properties:
              productId: { type: string, format: uuid }
              variantId: { type: string, format: uuid }
              batchNumber: { type: string, example: BATCH-SEP-2026 }
              expiryDate: { type: string, format: date }
              quantity: { type: string, example: '100' }
              unitCost: { type: string, example: '1050.00' }
    UpdatePurchaseRequest:
      type: object
      properties:
        supplierId: { type: string, format: uuid }
        purchaseNumber: { type: string }
        purchaseDate: { type: string, format: date }
        notes: { type: string }
        items:
          type: array
          items:
            type: object
            properties:
              productId: { type: string, format: uuid }
              quantity: { type: string }
              unitCost: { type: string }

    # ---- Billing ----
...